@@ -13,52 +13,52 @@
   "private": true,
   "dependencies": {
     "@angular-devkit/build-angular": "^0.803.21",
-    "@angular/animations": "7.1.3",
-    "@angular/cdk": "^7.0.3",
-    "@angular/common": "7.1.3",
-    "@angular/compiler": "7.1.3",
-    "@angular/core": "7.1.3",
-    "@angular/forms": "7.1.3",
-    "@angular/http": "7.1.3",
-    "@angular/material": "^7.0.3",
-    "@angular/platform-browser": "7.1.3",
-    "@angular/platform-browser-dynamic": "7.1.3",
-    "@angular/pwa": "^0.13.8",
-    "@angular/router": "7.1.3",
-    "@angular/service-worker": "7.1.3",
+    "@angular/animations": "8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/forms": "8.2.14",
+    "@angular/http": "7.2.15",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/pwa": "^0.803.21",
+    "@angular/router": "8.2.14",
+    "@angular/service-worker": "8.2.14",
     "angular2-uuid": "^1.1.1",
-    "angularfire2": "^5.0.0-rc.10",
-    "bootstrap": "^4.3.1",
-    "core-js": "^2.4.1",
-    "firebase": "5.1.0",
-    "mobx": "^5.8.0",
-    "mobx-angular": "^3.0.1",
-    "node-sass": "^4.5.3",
-    "npm": "^6.4.1",
-    "rxjs": "^6.3.3",
-    "rxjs-compat": "^6.3.3",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.8.26"
+    "angularfire2": "^5.2.3",
+    "bootstrap": "^4.4.1",
+    "core-js": "^3.6.1",
+    "firebase": "7.6.1",
+    "mobx": "^5.15.1",
+    "mobx-angular": "^3.0.3",
+    "node-sass": "^4.13.0",
+    "npm": "^6.13.4",
+    "rxjs": "^6.5.4",
+    "rxjs-compat": "^6.5.4",
+    "tslib": "^1.10.0",
+    "zone.js": "^0.10.2"
   },
   "devDependencies": {
     "@angular/cli": "^8.3.21",
-    "@angular/compiler-cli": "7.1.3",
-    "@angular/language-service": "7.1.3",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.2.0",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "3.1.6"
+    "@angular/compiler-cli": "8.2.14",
+    "@angular/language-service": "8.2.14",
+    "@types/jasmine": "~3.5.0",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/node": "~13.1.2",
+    "codelyzer": "~5.2.1",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-cli": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.5.1",
+    "protractor": "~5.4.2",
+    "ts-node": "~8.5.4",
+    "tslint": "~5.20.1",
+    "typescript": "^3.5.3"
   }
 }
